perf(p1b): use bfloat16 (MI300X native), batch_size=16 for 2x throughput

+++ baseline/eval_baseline.py
@@ -121,7 +121,7 @@ def main():
     tokenizer.pad_token = tokenizer.eos_token
     tokenizer.pad_token_id = tokenizer.eos_token_id
 
-    model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.float16)
+    model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.bfloat16)
     model.generation_config.enable_thinking = False
     model.eval()
     model.to("cuda" if torch.cuda.is_available() else "cpu")
